Move constants.ts to data/constants.ts
