+++ src/components/CountrySwitcher/DesktopCountrySwitcher.js
@@ -8,7 +8,7 @@ import {
   MenuButton as ReakitMenuButton,
 } from 'reakit/Menu'
 import LanguageContext from 'context/LanguageContext'
-import { LANGUAGES } from 'constants.ts'
+import { LANGUAGES } from 'data/constants'
 import Typography from 'components/Typography'
 
 const Container = styled.div`
