+++ src/components/CountrySwitcher/MobileCountrySwitcher.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 import LanguageContext from 'context/LanguageContext'
-import { LANGUAGES } from 'constants.ts'
+import { LANGUAGES } from 'data/constants'
 import Radio from 'components/Radio'
 import { useRadioState, RadioGroup } from 'reakit/Radio'
